+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Change Log
+## 1.18.0
+### Changes
+* `launch init` now updates the **package.json** file
+* `launch init` now supports templates
+* Added integration tests
+
 ## 1.17.3
 ### Changes
 * Added integration tests